requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
python-dotenv==1.0.1
langchain-openai==0.2.14
langgraph==0.2.58
python-multipart==0.0.12
mangum==0.18.0